add the `sort` option to help guarantee the same load order of `require()`d modules cross-platform.
comes with tests and README update.

Note: the tests are split into two directories `sort1` and `sort2` as unfortunately-for-our-tests any module which was `requireDir()`d once *stays loaded*, hence to check sorted vesus unsorted load order we have to apply either test to a different test directory or the test code (each module updates a global array when loaded) would fail on the second test.